Add MPS file parser (#123)
Adds the ability to read MPS files to create OMMX instances.

This implementation is based on our internal parser used in jijmodeling.
Some things could probably be changed to make the code simpler as they
were mostly choices to make it convenient/fast to create JijModeling
models specifically (eg. how the parser handles keeping track of
variable and constraint types), but nothing major was changed to get the
initial implementation done quicker.

This PR at the moment only includes the Rust interface for it, I'll be
commiting the python side with some tests later, but the code should be
reviewable at this point.

This adds flate2 and indoc as dependencies (indoc is only a
dev-dependency). flate2 is required to read files from miplib directly.
indoc is only for a minor test so let me know if it should be rewritten
to avoid the dependency