Adding Documentation for Downloading MIPLIB Instance (#169)
# Adding Documentation for Downloading Instances

This PR introduces a Japanese tutorial on how to download MIPLIB
instances using the OMMX Python SDK.

The tutorial deliberately avoids referencing or assuming prior knowledge
of the Container Registry. To align the page title with the content, I
renamed the placeholder page `pull_artifact` to the more descriptive
`download_miplib_instance`.