feat: adds `constraint_hints` field to instance and provide a scheme for one-hot constraints (#180)
This PR adds the following field to `ommx.v1.instance`:

- `constraint_hints`: A list of constraint hints to be used by solver to
gain performance. They are derived from one-or-more constraints in the
instance and typically contains information of special types of
constraints (e.g. one-hot, SOS, ...).

This is expressed as a `ConstraintHint` messag, which is a pair of
originating constraints and parameters specific to each special
constraint types.
Currently, there is only one constraint type supported: `OneHot`, which
expresses a one-hot constraint.

---------

Co-authored-by: Toshiki Teramura <toshiki.teramura@gmail.com>