MPS output is broken if same name variables exist (#162)
Fixes problems caused by trying to create and then read MPS files using
an instance that has multiple variables/constraints with the same name.

- When creating an MPS file, variables and constraints are now named
based on their IDs. The name metadata in the OMMX instance is ignored.
- When reading an OMMX file, if all names match the new ID-based naming
scheme, those IDs will be recovered and reused. Otherwise, the names are
interpreted as metadata and IDs are assigned in an unspecified order, as
before.

---------

Co-authored-by: Zengor <ialmeida@edu.unifor.br>