MPS: Only write used variables to output (#229)
As a lot of solvers error out if the `BOUNDS` section includes unused
variables, this PR updates the MPS output writer to only add variables
that were actually used in the problem to the `BOUNDS` section.

In this PR I also added a generic `anyhow::Error`→`MpsWriteError`
conversion asn "Unknown Error", but this is mostly just to make the code
compile. The method `used_variable_ids()` returns an `anyhow::Result`,
but it doesn't seem to be able to produce any errors at all. Still,
changing that method would be an API change so I decided to leave it as
is. I added the error conversion instead of unwrapping just in case.