Fix the index in the document of log encoding (#385)
This pull request includes updates to the release notes and
documentation to correct the formula for integer variable encoding. The
changes ensure consistency and accuracy across different language
versions and code documentation.

Documentation updates:

*
[`docs/en/release_note/ommx-1.9.0.ipynb`](diffhunk://#diff-542afa504675a8dc412a9b3fe943e47716afd9debc47969aa4ee918c7bc7a05aL33-R33):
Corrected the exponent in the integer variable encoding formula from
`2^l` to `2^i` to accurately reflect the intended calculation.
*
[`docs/ja/release_note/ommx-1.9.0.ipynb`](diffhunk://#diff-0efdfa43c848f444abf48f8ba007b97a2158a2c254a8141e2b94647adfebef2cL33-R33):
Made the same correction to the Japanese version of the release notes,
ensuring the formula uses `2^i` instead of `2^l`.

Code documentation update:

*
[`python/ommx/ommx/v1/__init__.py`](diffhunk://#diff-1bdad31bc2a9551bd1fb2f700436e715f09a483c70cdce58e8723541a855d52aL915-R915):
Updated the `log_encode` method's docstring to correct the exponent in
the formula, changing `2^l` to `2^i` for consistency with the corrected
documentation.